This is the base file
This R script creates a function, “cityQuery” that uses an input of a
list of cities + the designation of either “city” or “metro” to
generate a list of summary statistics for the cities included in the
list. After the function, several inputs are included that can be run,
in order, to generate results.